+++ CHANGELOG.md
@@ -43,6 +43,8 @@ BUG FIXES:
      Terraform structures. [GH-177]
   * core: Resources with only `file()` calls will interpolate. [GH-159]
   * core: Variables work in block names. [GH-234]
+  * core: Plugins are searched for in the same directory as the executable
+      before the PATH. [GH-157]
   * command/apply: "tfvars" file no longer interferes with plan apply. [GH-153]
   * providers/aws: Fix issues around failing to read EIPs. [GH-122]
   * providers/aws: Autoscaling groups now register and export load